+++ 2022Robot/src/main/java/frc/robot/commands/AutoRoutines/DoNothing.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutoRoutines;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
