+++ 2022Robot/src/main/java/frc/robot/commands/AutoRoutines/PickUpAndShoot.java
@@ -2,13 +2,17 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutoRoutines;
 
 import edu.wpi.first.math.estimator.DifferentialDrivePoseEstimator;
 import edu.wpi.first.wpilibj2.command.ParallelRaceGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.Constants; 
+import frc.robot.Constants;
+import frc.robot.commands.DeployIntake;
+import frc.robot.commands.Shoot;
+import frc.robot.commands.Suck;
+import frc.robot.commands.TimedDriveWithSpeed;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.subsystems.ShooterSubsystem;
