+++ 2022Robot/src/main/java/frc/robot/commands/AutoRoutines/ShootAndLeave.java
@@ -2,22 +2,28 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutoRoutines;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import edu.wpi.first.wpilibj2.command.WaitCommand;
+import frc.robot.Constants;
+import frc.robot.commands.Shoot;
+import frc.robot.commands.TimedDriveWithSpeed;
 import frc.robot.subsystems.DriveSubsystem;
+import frc.robot.subsystems.ShooterSubsystem;
+import frc.robot.subsystems.TransferSubsystem;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class LeaveTarmac extends SequentialCommandGroup {
-  /** Creates a new LeaveTarmac. */
-  public LeaveTarmac(DriveSubsystem driveSubsystem) {
+public class ShootAndLeave extends SequentialCommandGroup {
+  /** Creates a new ShootAndLeave. */
+  public ShootAndLeave(TransferSubsystem transferSubsystem, ShooterSubsystem shooterSubsystem, DriveSubsystem driveSubsystem) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
-      new TimedDriveWithSpeed(0.5, 1, driveSubsystem)
+      new Shoot(transferSubsystem, shooterSubsystem, Constants.ShooterConstants.kLowSpeed).withTimeout(4),
+
+      new TimedDriveWithSpeed(0.5, 1.5, driveSubsystem)
     );
   }
 }
