+++ 2022Robot/src/main/java/frc/robot/commands/AutoRoutines/TraverseRungs.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.AutoRoutines;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ClimbSubsystem;
