+++ 2022Robot/src/main/java/frc/robot/commands/Simulation/DriveToDistance.java
@@ -2,36 +2,37 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Simulation;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.wpilibj2.command.PIDCommand;
-import frc.robot.subsystems.ClimbSubsystem;
+import frc.robot.subsystems.DriveSubsystem;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class SetElevatorToHeightPID extends PIDCommand {
-  private ClimbSubsystem m_climbSubsystem;
-  /** Creates a new ExtendElevatorToHeight. */
-  public SetElevatorToHeightPID(double setHeight, ClimbSubsystem climbSubsystem) {
+public class DriveToDistance extends PIDCommand {
+
+  private final DriveSubsystem m_driveSubsystem;
+  /** Creates a new DriveToDistance. */
+  public DriveToDistance(double targetDist, DriveSubsystem driveSubsystem) {
     super(
         // The controller that the command will use
-        new PIDController(0.5, 0, 0),
+        new PIDController(1, 0, 0),
         // This should return the measurement
-        () -> climbSubsystem.getElevatorRotations(),
+        driveSubsystem::getAvgEncoderDistance,
         // This should return the setpoint (can also be a constant)
-        setHeight,
+        targetDist,
         // This uses the output
         output -> {
           // Use the output here
-          climbSubsystem.setElevator(output);
-        });
-
-    m_climbSubsystem = climbSubsystem;
-    getController().setTolerance(1, 0.1);
+          driveSubsystem.driveRaw(output, output);
+        }, driveSubsystem);
+    m_driveSubsystem = driveSubsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     // Configure additional PID options by calling `getController` here.
+
+    getController().setTolerance(1, 0.1);
   }
 
   // Returns true when the command should end.
@@ -42,6 +43,8 @@ public class SetElevatorToHeightPID extends PIDCommand {
 
   @Override
   public void end(boolean interrupted){
-    m_climbSubsystem.stopElevator();
+    // m_driveSubsystem.refreshOdometry();
+    m_driveSubsystem.resetEncoders();
+    // m_driveSubsystem.resetEncoderSims();
   }
 }
