+++ 2022Robot/src/main/java/frc/robot/commands/Simulation/SetElevatorToHeightPID.java
@@ -2,35 +2,36 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Simulation;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.wpilibj2.command.PIDCommand;
-import frc.robot.subsystems.DriveSubsystem;
+import frc.robot.subsystems.ClimbSubsystem;
 
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class TurnToAngle extends PIDCommand {
-  /** Creates a new TurnToAngle. */
-  public TurnToAngle(double targetAngleDeg, DriveSubsystem driveSubsystem) {
+public class SetElevatorToHeightPID extends PIDCommand {
+  private ClimbSubsystem m_climbSubsystem;
+  /** Creates a new ExtendElevatorToHeight. */
+  public SetElevatorToHeightPID(double setHeight, ClimbSubsystem climbSubsystem) {
     super(
         // The controller that the command will use
-        new PIDController(0.008, 0, 0),
+        new PIDController(0.5, 0, 0),
         // This should return the measurement
-        driveSubsystem::getHeading,
+        () -> climbSubsystem.getElevatorRotations(),
         // This should return the setpoint (can also be a constant)
-        targetAngleDeg,
+        setHeight,
         // This uses the output
         output -> {
           // Use the output here
-          driveSubsystem.driveArcade(0, output);
-        }, driveSubsystem);
+          climbSubsystem.setElevator(output);
+        });
+
+    m_climbSubsystem = climbSubsystem;
+    getController().setTolerance(1, 0.1);
     // Use addRequirements() here to declare subsystem dependencies.
     // Configure additional PID options by calling `getController` here.
-
-    getController().enableContinuousInput(0, 360);
-    getController().setTolerance(2, 1);
   }
 
   // Returns true when the command should end.
@@ -39,5 +40,8 @@ public class TurnToAngle extends PIDCommand {
     return getController().atSetpoint();
   }
 
-  
+  @Override
+  public void end(boolean interrupted){
+    m_climbSubsystem.stopElevator();
+  }
 }
