+++ 2022Robot/src/main/java/frc/robot/commands/Simulation/TurnToAngle.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Simulation;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.wpilibj2.command.PIDCommand;
@@ -11,28 +11,26 @@ import frc.robot.subsystems.DriveSubsystem;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class DriveToDistance extends PIDCommand {
-
-  private final DriveSubsystem m_driveSubsystem;
-  /** Creates a new DriveToDistance. */
-  public DriveToDistance(double targetDist, DriveSubsystem driveSubsystem) {
+public class TurnToAngle extends PIDCommand {
+  /** Creates a new TurnToAngle. */
+  public TurnToAngle(double targetAngleDeg, DriveSubsystem driveSubsystem) {
     super(
         // The controller that the command will use
-        new PIDController(1, 0, 0),
+        new PIDController(0.008, 0, 0),
         // This should return the measurement
-        driveSubsystem::getAvgEncoderDistance,
+        driveSubsystem::getHeading,
         // This should return the setpoint (can also be a constant)
-        targetDist,
+        targetAngleDeg,
         // This uses the output
         output -> {
           // Use the output here
-          driveSubsystem.driveRaw(output, output);
+          driveSubsystem.driveArcade(0, output);
         }, driveSubsystem);
-    m_driveSubsystem = driveSubsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     // Configure additional PID options by calling `getController` here.
 
-    getController().setTolerance(1, 0.1);
+    getController().enableContinuousInput(0, 360);
+    getController().setTolerance(2, 1);
   }
 
   // Returns true when the command should end.
@@ -41,10 +39,5 @@ public class DriveToDistance extends PIDCommand {
     return getController().atSetpoint();
   }
 
-  @Override
-  public void end(boolean interrupted){
-    // m_driveSubsystem.refreshOdometry();
-    m_driveSubsystem.resetEncoders();
-    // m_driveSubsystem.resetEncoderSims();
-  }
+  
 }
